Create new scripts and added new system
I created a Special attack system for the green candy. Which checks everytime a match is made with the green candy. If that happens it performs a special attack. I also checks if the special bar is full before performing a normal attack to prevent doing double damage. also changed the selection screen icons. Had to make changes in the gridmanager script to get everything working together with the newly added scripts which are the GreenSpecialController and SpecialBar.